Changed sample application to avoid Tcl error when destroying Button drag source.

+++ samples/hello/hello_drag_and_drop.rb
@@ -126,11 +126,11 @@ root {
         grid :row => 4, :column => 0
       }
       checkbutton {
-        text "Check"
+        text "Drop here to destroy a widget\n(except button)"
         grid :row => 4, :column => 1, :pady => 5, :sticky => "w"
         on_drop { |event|
           event.target.text = event.data
-          event.source.destroy
+          event.source.destroy unless event.source.is_a? Tk::Button
         }
       }
     }
